Makes link ignore symlinks
There are many different ways a symlink can be handled and it is not currently clear to
me what is the best way.

> In the face of ambiguity, refuse the temptation to guess.

Ignoring them allows the user to manually fill in the missing pieces as they see fit.
It also makes it easier to add symlink handling in the future since there will likely
be fewer conflicts between the symlink we want to write to and the symlink that already
exists in the destination.