Makes link detect all conflicts
The command should strive to have no effect if it fails. Previously it would not detect
conflicts with directories or broken links early and fail later when it had may already
have created a bunch of links.

This is sort of an ask-for-permission approach, the ask-for-forgiveness would try to
clean up its effects on failure but that
* is harder,
* is dangerous since the celan up could potentially be destructive, and
* is harder still if the action itself was destructive.