Changes link interface removing include option
As a required option this made usage less ergonomic as the typical use case (for me) is
mirroring all files under a directory. The globbing was complex and probably buggy so
rather than making it optional it was removed entirely.

The reason for having three parameters originally was that the dst parameter also would
demarcate the boundary inside which relative symlinks would be preserved. Preserving
symlinks is no longer a priority and when I do I think detecting the root or providing
it explicitly would be more ergonomic and allow placing links at a path inside the
boundary that perhaps does not exist in the src.