Rewrites tests
Refactors workflow steps as individual tests to eliminate need for manual clean up and
make it easier to see what parts are failing.

Removes testing of both modified index entry and modified file content as these are
equivalent.

Adds tests for check failing on missing index file, as would happen if one adds a link
but forgets to track it.

Adds tests for check succeeding on paths not related to the invalid link/index. The h
file is added to make sure corrupting g does not affect the validity of any siblings.

Splits tests for link to check that it would succeed if existing links are the same as
the ones it tried to create, as would happen if one tries to fill in a directory with
new data.

Checks that functions are nullipotent where appropriate