Makes commands non-destructive by default
Because I think it is nice UX to stop unfamiliar commands doing something that the user
does not intend and that cannot be undone.

The plan is to implement
* r (dry-run) mode for all commands
* cr (normal) mode for track and
* cru (force) mode for link

The interface will probably be revamped at some point. I would have liked to use the
somewhat conventional -n and -f but it seems it is not possible to specify
single-hyphen flags with fire. I will have to think of something.