Makes link work if destination exists
Will fill in missing links, skip existing if they are the same and panic if a link
exists in the destination but does not point to the same source.

This should make it easier to incrementally link a data collections as it builds up
while maintaining the safety and confidence that comes from having a
non-destructive command.