Replaces the crud option
The new option is inspired by the `-X` option that is used with the `--strategy` option
in some git commands. It is a smaller interface (easier to understand and test), can be
omitted most of the time (more ergonomic), allows the equivalent of a force mode, and
allow commands like link to ignore existing files and continue or stop.

Can not be used to achieve dry-run behavior.

Could potentially be used also for an interactive resolution mode.