Replaces opaque fingerprint function
The property of the previous version that the size of the fingerprint is constant in
the number of and size of nodes is not important since the tests operate on small trees
with tiny content.

On the other hand, if the fingerprints where different, the diff did not provide any
information about why.