Make directory fingerprint more robust and easier to read
Easier to read because we now see what file or subdirectory correspond to what
attributes and differences were both the number of items and their attributes have
changed show what is new and what is changed.

More robust because the order in which items are returned from `iterdir` is not well
defined and without a name one could in theory rename an item and still get the same
fingerprint.